change a few packages
