+++ xmlunit-jakarta-jaxb-impl/src/main/resources/META-INF/services/org.xmlunit.builder.JaxbBuilderFactory
@@ -1,1 +1,1 @@
-org.xmlunit.jakarta_jaxb.JakartaJaxbBuilderFactory
+org.xmlunit.builder.jakarta_jaxb.JakartaJaxbBuilderFactory
